Rename package, remove egg-info
