+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## 0.2.0
+
+## Changed
+- Renamed package from `pygeotemporal` to `pygeostreams` for clarity.
+## 0.1.0
 
 ### Added 
 - Added timeout to requests
