+++ docs/source/conf.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Pygeotemporal documentation build configuration file, created by
+# Pygeostreams documentation build configuration file, created by
 # sphinx-quickstart on Tue Jan 24 09:57:45 2017.
 #
 # This file is execfile()d with the current directory set to its
@@ -19,7 +19,7 @@
 import os
 import sys
 sys.path.insert(0, os.path.abspath('../..'))
-sys.path.insert(0, os.path.abspath('../../pygeotemporal'))
+sys.path.insert(0, os.path.abspath('../../pygeostreams'))
 
 # -- General configuration ------------------------------------------------
 
@@ -48,7 +48,7 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = u'Pygeotemporal'
+project = u'Pygeostreams'
 copyright = u'2017, Luigi Marini, Yan Zhao, Marcus Slavenas'
 author = u'Luigi Marini, Yan Zhao, Marcus Slavenas'
 
@@ -100,7 +100,7 @@ html_static_path = ['_static']
 # -- Options for HTMLHelp output ------------------------------------------
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'Pygeotemporaldoc'
+htmlhelp_basename = 'Pygeostreamsdoc'
 
 # -- Options for LaTeX output ---------------------------------------------
 
@@ -126,7 +126,7 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'Pygeotemporal.tex', u'Pygeotemporal Documentation',
+    (master_doc, 'Pygeostreams.tex', u'Pygeostreams Documentation',
      u'Luigi Marini, Yan Zhao, Marcus Slavenas', 'manual'),
 ]
 
@@ -135,7 +135,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'Pygeotemporal', u'Pygeotemporal Documentation',
+    (master_doc, 'Pygeostreams', u'Pygeostreams Documentation',
      [author], 1)
 ]
 
@@ -145,7 +145,7 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'Pygeotemporal', u'Pygeotemporal Documentation',
-     author, 'Pygeotemporal', 'One line description of project.',
+    (master_doc, 'Pygeostreams', u'Pygeostreams Documentation',
+     author, 'Pygeostreams', 'One line description of project.',
      'Miscellaneous'),
 ]
