+++ pygeostreams/__init__.py
@@ -1,0 +1,4 @@
+from pygeostreams.client import GeostreamsClient
+from pygeostreams.sensors import SensorsApi
+from pygeostreams.streams import StreamsApi
+from pygeostreams.datapoints import DatapointsApi
