+++ pygeostreams/load_csv.py
@@ -8,7 +8,7 @@ from builtins import range
 from builtins import object
 import logging
 
-from pygeotemporal.time_transformers import time2utc
+from pygeostreams.time_transformers import time2utc
 
 
 def load(file):
