+++ pygeostreams/sensors.py
@@ -11,7 +11,7 @@ from typing import Union, Tuple
 
 from requests.exceptions import RequestException
 
-from pygeotemporal.client import GeostreamsClient
+from pygeostreams.client import GeostreamsClient
 
 
 class SensorsApi(object):
