--- pygeotemporal.egg-info/PKG-INFO
@@ -1,10 +1,0 @@
-Metadata-Version: 1.0
-Name: pygeotemporal
-Version: 0.1.0
-Summary: UNKNOWN
-Home-page: UNKNOWN
-Author: Luigi Marini, Yan Zhao, Marcus Slavenas
-Author-email: lmarini@illinois.edu
-License: NCSA
-Description: UNKNOWN
-Platform: UNKNOWN
