--- pygeotemporal.egg-info/SOURCES.txt
@@ -1,23 +1,0 @@
-setup.cfg
-pygeotemporal/__init__.py
-pygeotemporal/client.py
-pygeotemporal/collections.py
-pygeotemporal/connectors.py
-pygeotemporal/datasets.py
-pygeotemporal/extractors.py
-pygeotemporal/files.py
-pygeotemporal/sections.py
-pygeotemporal/utils.py
-pygeotemporal.egg-info/PKG-INFO
-pygeotemporal.egg-info/SOURCES.txt
-pygeotemporal.egg-info/dependency_links.txt
-pygeotemporal.egg-info/top_level.txt
-pygeotemporal/geostreams/__init__.py
-pygeotemporal/geostreams/cache.py
-pygeotemporal/geostreams/csv.py
-pygeotemporal/geostreams/datapoints.py
-pygeotemporal/geostreams/datasets.py
-pygeotemporal/geostreams/map_names.py
-pygeotemporal/geostreams/sensors.py
-pygeotemporal/geostreams/streams.py
-pygeotemporal/geostreams/time_transformers.py
