--- pygeotemporal/__init__.py
@@ -1,4 +1,0 @@
-from pygeotemporal.client import GeostreamsClient
-from pygeotemporal.sensors import SensorsApi
-from pygeotemporal.streams import StreamsApi
-from pygeotemporal.datapoints import DatapointsApi
